move project setup docs to testing docs page
